scripts/scalping/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "scalping"
version = "0.1.0"
description = "Scalping variant of the BloFin TradingView webhook bridge (cloned from blofin-bridge)"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "ccxt==4.4.50",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "httpx==0.27.2",
    "PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-mock==3.14.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
